feat: Implement contents toggle (#88)
* style: Fix font size

* refactor: Implement PostControl component

* chore: Add todo

* chore: Add toggle button

* feat: Implement toggle

* refactor: Remove magic number

* refactor: Separate Post contents component

* refactor: Move handleDelete to PostControl component
