+++ src/components/Post.tsx
@@ -1,18 +1,17 @@
-import { Link, useNavigate } from "react-router-dom";
-import { FormEvent, useState } from "react";
+import { useState } from "react";
 import { BoxArrowUpRight } from "react-bootstrap-icons";
 import { IPost } from "../@types/posts";
 import Carousel from "./Carousel";
-import Button from "./Button";
 import useAuthStore from "../store/useAuthStore";
-import { addLike, deletePost, removeLike } from "../api/posts";
+import { addLike, removeLike } from "../api/posts";
 import toast from "../utils/toast";
-import "./Post.css";
 import { Heart } from "./Heart";
+import { PostControl } from "./PostControl";
+import "./Post.css";
+import { PostContents } from "./PostContents";
 
 export default function Post({ post }: { post: IPost }) {
   const { user } = useAuthStore();
-  const go = useNavigate();
 
   const { id, title, contents, address, created, images, beliked, likes } =
     post;
@@ -51,45 +50,12 @@ export default function Post({ post }: { post: IPost }) {
     toast("Copied it!");
   };
 
-  const handleDelete = async (event: FormEvent) => {
-    event.preventDefault();
-
-    if (!confirm("정말 삭제하시겠습니까?")) {
-      return;
-    }
-
-    const { response, data } = await deletePost(id);
-
-    if (!response.ok) {
-      toast(data.message);
-      return;
-    }
-
-    go("/");
-  };
-
   return (
     <div className="post">
       <Carousel title={title} images={images} />
 
       <article className="post__description">
-        {user?.role === "ADMIN" ? (
-          <div className="post__control">
-            <Link to={`/posts/update/${id}`} state={post}>
-              <Button value="수정하기" size="small" />
-            </Link>
-            <form method="DETELE" onSubmit={handleDelete}>
-              <Button
-                type="submit"
-                value="삭제하기"
-                btnStyle="secondary"
-                size="small"
-              />
-            </form>
-          </div>
-        ) : (
-          ""
-        )}
+        {user?.role === "ADMIN" ? <PostControl id={id} post={post} /> : ""}
         <h1 className="post__title">{title}</h1>
         <div className="post__date">{date}</div>
         <div className="post__action">
@@ -101,7 +67,7 @@ export default function Post({ post }: { post: IPost }) {
           <BoxArrowUpRight className="post__link" onClick={handleClickLink} />
         </div>
         <div className="post__address">{address}</div>
-        <div className="post__contents">{contents}</div>
+        <PostContents contents={contents} />
       </article>
     </div>
   );
