+++ src/components/PostControl.css
@@ -1,0 +1,4 @@
+.post__control {
+  display: flex;
+  justify-content: flex-end;
+}
